+++ scout/commands/view/panels.py
@@ -3,6 +3,7 @@ import logging
 import click
 from flask.cli import with_appcontext
 
+from scout.constants import DATE_DAY_FORMATTER
 from scout.server.extensions import store
 
 LOG = logging.getLogger(__name__)
@@ -31,6 +32,6 @@ def panels(institute):
                 str(panel_obj["version"]),
                 len(panel_obj["genes"]),
                 panel_obj.get("hidden", False),
-                str(panel_obj["date"].strftime("%Y-%m-%d")),
+                str(panel_obj["date"].strftime(DATE_DAY_FORMATTER)),
             )
         )